umami_service/pyproject.toml
[tool.poetry]
name = "umami_service"
version = "0.1.0"
description = "umami service"
authors = ["petryanin <aspetryanin@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.5"
fastapi = "^0.103.0"
SQLAlchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
colorlog = "^6.7.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
alembic = "^1.12.0"
asyncpg = "^0.28.0"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"
